% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{AWR.Lambda}
\alias{AWR.Lambda}
\alias{AWR.Lambda-package}
\title{An R client to Amazon Lambda}
\description{
This is a simple wrapper around the most important feature of the related Java SDK, \code{invoke()}.
}
\references{
\url{http://docs.aws.amazon.com/AWSJavaSDK/latest/javadoc/com/amazonaws/services/lambda/AWSLambdaClient.html}
}
