% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{invoke}
\alias{invoke}
\alias{lambda_client}
\title{Invoke a Lambda function}
\usage{
invoke(name, payload, ..., context, client = lambda_client())

lambda_client()
}
\arguments{
\item{name}{name of the lambda function}

\item{payload}{a toJSON-able R object}

\item{...}{not used}

\item{context}{the client context}

\item{client}{a Lambda client (invoke will be called upon it)}
}
\value{
function result
}
\description{
Invoke a Lambda function
}
\examples{
\dontrun{
# Using the default hello-world lambda function, default event, and default auth:
AWR.Lambda::invoke("myFunctionName",  list(key3="value3", key2="value2",key1="value1"))
}
}
